# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YAWL REST API
  description: |
    RESTful API for YAWL (Yet Another Workflow Language) v5.2 Workflow Engine

    This API provides comprehensive workflow management capabilities including:
    - Session management and authentication
    - Work item lifecycle operations (checkout, checkin, complete)
    - Case management and control
    - Specification management and validation
    - Event subscriptions and monitoring
    - Extended operations (exception handling, task suspension/resumption)

    All operations require a valid session handle obtained through the `/ib/connect` endpoint.
  version: 5.2.0
  contact:
    name: YAWL Foundation
    url: https://www.yawlfoundation.org
  license:
    name: GNU Lesser General Public License v2.1
    url: https://www.gnu.org/licenses/old-licenses/lgpl-2.1.html
servers:
  - url: http://localhost:8080/yawl/api
    description: Local development server
    variables:
      port:
        default: '8080'
  - url: https://yawl.example.com/api
    description: Production server
...
paths:
  /ib/connect:
    post:
      tags:
        - Session Management
      summary: Connect to the YAWL engine
      description: |
        Authenticates a user and establishes a session with the engine.
        Returns a session handle that must be used for subsequent operations.
      operationId: connect
      requestBody:
        required: true
        description: User credentials for authentication
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            examples:
              basic:
                summary: Basic authentication
                value:
                  userid: admin
                  password: admin
      responses:
        '200':
          description: Successful connection - returns session handle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              examples:
                success:
                  value:
                    sessionHandle: "H4sIAAAAAAAAAKtWSkksSVSyUkopLUsB0oosSixJVLJSSiiNL8nPzy_JBKkpqMxLySwBsVNSUkpRSWIVoFNKUUlqUWpxZklJZkZqUWpxSWZxcWZRSklOTklKZm5ySmZiVmZOSlFiUmlyRnFxSklRUWpRSmZxSWZRSWVBeXlBUXZCUGVRYXVKZnFGZl5aZl1iUmlyRGlyRklySkllZkJOVoFGQWZxZklmcWlxSV5RZmlyTmZJZnFqUUVRZXFBSV1+SWlxSWZRSWZJUlFiSWlSaHF+SUpKZX1NZUlqUWlyCUVGWSmpxSWZxcWlySk1mSSllSWpOJUVFqSX5+YUlySllQWZ1QmZxZmpJZn5BaklmSXFBZmlyYklKZnFxZkliSWlxSXlJS--9OLS0KDgxJBKsAgNg89CyAAAAA"
        '401':
          description: Authentication failed - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  value:
                    error: "Invalid credentials"
        '400':
          description: Bad request - missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_fields:
                  value:
                    error: "Missing userid or password"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /ib/workitems/{itemId}/checkin:
    post:
      tags:
        - Work Items
      summary: Check in a work item
      description: |
        Updates work item data without completing it.
        Useful for saving intermediate progress.
      operationId: checkinWorkItem
      parameters:
        - name: itemId
          in: path
          required: true
          description: Work item ID
          schema:
            type: string
        - name: sessionHandle
          in: query
          required: true
          description: Valid session handle
          schema:
            type: string
      requestBody:
        required: true
        description: Work item data in XML format
        content:
          application/xml:
            schema:
              type: string
      responses:
        '200':
          description: Successful check-in
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "checked-in"
        '400':
          description: Failed to checkin work item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing session handle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Credentials:
      type: object
      required:
        - userid
        - password
      properties:
        userid:
          type: string
          description: Username for authentication
          example: admin
        password:
          type: string
          format: password
          description: User password
          example: admin
    SessionResponse:
      type: object
      required:
        - sessionHandle
      properties:
        sessionHandle:
          type: string
          description: |
            Session handle for use in subsequent requests.
            Must be included as a query parameter in all operations.
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing the failure
...
  securitySchemes:
    SessionHandle:
      type: apiKey
      in: query
      name: sessionHandle
      description: Session handle obtained from /ib/connect endpoint
security:
  - SessionHandle: []
...